+++ lib/msal-core/lib-es6/Logger.d.ts
@@ -1,25 +1,3 @@
-/**
- * Copyright (c) Microsoft Corporation
- *  All Rights Reserved
- *  MIT License
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this
- * software and associated documentation files (the 'Software'), to deal in the Software
- * without restriction, including without limitation the rights to use, copy, modify,
- * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to the following
- * conditions:
- *
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED 'AS IS', WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS
- * OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
- * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT
- * OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
 export interface ILoggerCallback {
     (level: LogLevel, message: string, containsPii: boolean): void;
 }
@@ -27,7 +5,7 @@ export declare enum LogLevel {
     Error = 0,
     Warning = 1,
     Info = 2,
-    Verbose = 3
+    Verbose = 3,
 }
 export declare class Logger {
     private static _instance;
@@ -40,7 +18,7 @@ export declare class Logger {
         level?: LogLevel;
         piiLoggingEnabled?: boolean;
     });
-    private logMessage;
+    private logMessage(logLevel, logMessage, containsPii);
     executeCallback(level: LogLevel, message: string, containsPii: boolean): void;
     error(message: string): void;
     errorPii(message: string): void;